10-agent-examples/requirements.txt
# LangGraph 工作流所需依赖（最新稳定版，截至 2025-10-17）

# 核心
langgraph>=1.0.5
langchain>=1.2.0
langchain-core>=1.2.2

# OpenAI 提供商（使用 OpenAI API 时需要）
langchain-openai>=1.1.5

# 本地模型（使用 Ollama 时需要）
langchain-ollama>=1.0.1

# 社区集成
langchain-community>=0.4.1

# 常用辅助
tiktoken>=0.12.0
python-dotenv>=1.2.1

# 增强版 Agent 所需依赖
duckduckgo-search>=8.1.1 # Web 搜索工具
langchain-experimental>=0.4.1 # Python REPL 等实验性工具
